kafaa_oe_assessment_v25_3_full/requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
pyyaml==6.0.1
python-pptx==1.0.2
reportlab==4.1.0
kaleido==0.2.1
openpyxl==3.1.5
Pillow==10.3.0
